Error when non-errors are thrown or rejected
This adds two new rules to the default set:

  - [`no-throw-literal`](https://eslint.org/docs/latest/rules/no-throw-literal)
  - [`prefer-promise-reject-errors`](https://eslint.org/docs/latest/rules/prefer-promise-reject-errors)

We've already seen some very difficult-to-debug problems with throwing
non-errors in the newsletters page, because anything else that's thrown
will not result in an error stack being generated. We don't know the
scope of the issue but I think we should be more strict about throwing
error objects going forward.

We go into more detail about this issue in the [Reliability Kit docs](https://github.com/Financial-Times/dotcom-reliability-kit/blob/main/docs/getting-started/throwing-errors.md#using-error-objects).

If we decide to accept this change (I'm open to arguments as to why we
shouldn't) then I think this would be considered a breaking change to
the ESLint config, as some apps will fail to build when they upgrade.